producer/requirements.txt
kafka-python==2.3.0
pandas==2.3.3
python-dotenv==1.2.1
